chore: fix access to applocalization
- Enable access to an instance of AppLocalization from the test file
- Rename the onboarding screen integration test file
