+++ integration_test/app_test.dart
@@ -2,7 +2,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
 import 'bottom_nav_bar_test.dart';
-import 'onboarding/view/onboarding_screens.dart';
+import 'onboarding/view/onboarding_flow_test.dart';
+
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
